fix: fixed merge conflict and added a test script to test both package at once
