+++ package.json
@@ -7,6 +7,7 @@
   ],
   "scripts": {
     "build": "turbo run build",
+    "test":  "turbo test",
     "dev": "turbo run dev",
     "lint": "turbo run lint",
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
