+++ packages/contracts/package.json
@@ -4,6 +4,10 @@
   "main": "index.js",
   "license": "MIT",
   "packageManager": "yarn@3.5.0",
+  "scripts": {
+    "test": "hardhat test",
+    "coverage": "hardhat coverage"
+  },
   "dependencies": {
     "@gnosis.pm/safe-contracts": "^1.3.0",
     "@nomicfoundation/hardhat-chai-matchers": "^1.0.0",
